requirements.txt
fastapi
uvicorn[standard]
paho-mqtt
pydantic
mysql-connector-python
python-dotenv